// src/components/providers/b2b_slots/enums/enums.ts
export enum B2BCallbacks {
    AUTH = 'do-auth-user-ingame',
    DEBIT = 'do-debit-user-ingame',
    CREDIT = 'do-credit-user-ingame',
    ROLLBACK = 'do-rollback-user-ingame',
    GET_FEATURES = 'do-get-features-user-ingame',
    ACTIVATE_FEATURES = 'do-activate-features-user-ingame',
    END_FEATURES = 'do-end-features-user-ingame'
}