// src/components/casino_category/enums/casino_category.enum.ts
export enum CasinoCategoryType {
    TABLE_GAMES = 3,
    SLOTS = 4,
    VIDEO_POKER = 5,
    KENO = 6,
    SCRATCH_CARD = 7,
    BACCARAT = 8,
    BLACKJACK = 9,
    POKER = 10,
    VIDEOBINGOS = 11,
    LOTTERY = 12,
    MULTIHAND_POKER = 13,
    SOFT_GAMES = 14,
    PYRAMID_POKER = 15,
    ROULETTE = 16
}